Arguments and indList changes
System takes in an argument from the command line for what file to read and fails if incorrect usage. IndList will remove the most recently added ID if the name cannot be added. A counter now checks if at capacity and only increments after a successful insert of both ID and Name.